# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CrossWithFriends API",
    "description": "API for the CrossWithFriends crossword puzzle platform",
    "version": "1.0.0",
    "contact": {"name": "CrossWithFriends"}
  },
...
  "paths": {
...
    "/link_preview": {
      "get": {
        "operationId": "getLinkPreview",
        "summary": "Get link preview",
        "tags": ["Link Preview"],
        "description": "Generates Open Graph metadata for social media link previews",
        "parameters": [
          {
            "schema": {"type": "string"},
            "in": "query",
            "name": "url",
            "required": true,
            "description": "URL to generate preview for"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML with Open Graph metadata",
            "content": {
              "application/json": {
                "schema": {"type": "string", "description": "HTML with Open Graph metadata"}
              }
            }
          },
          "302": {"description": "Redirect for non-bot user agents"},
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "message"],
                  "properties": {
                    "error": {"type": "string", "description": "Error type"},
                    "message": {"type": "string", "description": "Error message"},
                    "statusCode": {"type": "number", "description": "HTTP status code"},
                    "retryAfter": {"type": "number", "description": "Retry after seconds (for rate limiting)"}
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "message"],
                  "properties": {
                    "error": {"type": "string", "description": "Error type"},
                    "message": {"type": "string", "description": "Error message"},
                    "statusCode": {"type": "number", "description": "HTTP status code"},
                    "retryAfter": {"type": "number", "description": "Retry after seconds (for rate limiting)"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {"url": "https://www.crosswithfriends.com/api", "description": "Production server"},
    {"url": "https://crosswithfriendsbackend-staging.onrender.com/api", "description": "Staging server"},
    {"url": "http://localhost:3021/api", "description": "Local development server"}
  ],
...
}